requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Sock==0.6.0
openai-whisper
soundfile
gunicorn
python-dotenv
torch
torchaudio
numpy
ffmpeg-python
